+++ CHANGES.md
@@ -5,6 +5,7 @@
 * PR #1494: Added Support for Change Data Capture (CDC)
 * Issue #1501: Preserved detailed row errors from failed direct writes
 * Issue #1503: Corrected the documented billing project option to `parentProject`.
+* PR #1510: Handle Arrow StructVectors with 0 child vectors in ArrowColumnBatchPartitionReaderContext
 
 ## 0.44.2 - 2026-05-20